Updated smoke test
