+++ lib/widgets/component/quick_menu.dart
@@ -1,6 +1,6 @@
+import 'package:concordia_go/utilities/application_constants.dart' as application_constants;
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:concordia_go/utilities/application_constants.dart' as application_constants;
 
 class QuickMenu extends StatefulWidget {
   @override
